completed the api for notes

+++ api/v1/views/vitals.py
@@ -1,61 +1,86 @@
 #!/usr/bin/python3
 """ objects that handle all default RestFul API actions for Vital Signs """
+from flasgger.utils import swag_from
+from flask import abort, jsonify, make_response, request
+
 from models.notes.vitals import VitalSign
 from storage import storage
 from api.v1.views import app_views
-from flask import abort, jsonify, make_response, request
-from flasgger.utils import swag_from
 
 
-@app_views.route('/patients/<pid>/vitals',
-                 methods=['GET'], strict_slashes=False)
+@app_views.route('/patients/<int:pid>/vitals', methods=['GET'],
+                 strict_slashes=False)
 @swag_from('documentation/patient/patient_id/get_vitals.yml')
 def get_vitals(pid):
-    """ Retrieves the all VitalSign object for a specific patient """
-    all_vitals = storage.all(VitalSign).values()
-    list_vitals = []
-    for vital in all_vitals:
-        if vital.pid == pid:
-            list_vitals.append(vital.to_dict())
-    return jsonify(list_vitals)
+    """
+    Retrieves the all VitalSign object for a specific patient
+    """
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
 
-@app_views.route('/patients/<pid>/consultations/<consultation_id>/vitals',
+    vitals = [vitals.to_dict() for vitals in patient.vitals]
+    return jsonify(vitals)
+
+
+@app_views.route('/patients/<int:pid>/consultations/<consultation_id>/vitals',
                  methods=['GET'], strict_slashes=False)
 @swag_from('documentation/patients/patient_id/consultations/consultation_id/get_vital.yml')
-def get_vital(consultation_id):
-    """ Retrieves the all vital object for a specific consultation """
-    all_vitals = storage.all(VitalSign).values()
-    list_vitals = []
-    for vital in all_vitals:
-        if vital.consultation_id == consultation_id:
-            list_vitals.append(vital.to_dict())
-    return jsonify(list_vitals)
+def get_consult_vitals(pid, consultation_id):
+    """
+    Retrieves the all vital object for a specific consultation
+    """
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
 
-@app_views.route('/patients/<pid>/consultations/<consultation_id>/vitals/<vital_id>',
+    consultation = storage.get('Consultation', 'pid', pid)
+    if (not consultation) or (consultation.pid != pid):
+        abort(404, description="Consultation not found")
+
+    vitals = [vitals.to_dict() for vitals in consultation.vitals]
+    return jsonify(vitals)
+
+
+@app_views.route('/patients/<int:pid>/consultations/<consultation_id>/vitals/<vital_id>',
                  methods=['GET'], strict_slashes=False)
 @swag_from('documentation/patients/patient_id/consultations/consultation_id/vitals/get_vital.yml')
-def get_vital(vital_id):
+def get_vital(pid, consultation_id, vital_id):
     """ Retrieves the a specific vital object for a specific consultation """
-    vital = storage.get(VitalSign, vital_id)
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
+
+    consultation = storage.get('Consultation', 'id', consultation_id)
+    if (not consultation) or (consultation.pid != pid):
+        abort(404, description="Consultation not found")
+
+    vital = storage.get('VitalSign', 'id', vital_id)
     if not vital:
-        abort(404)
+        abort(404, description="No such vitals")
 
     return jsonify(vital.to_dict())
 
-@app_views.route('/patients/<pid>/consultations/<consultation_id>/vitals/<vitals_id>',
+
+@app_views.route('/patients/<int:pid>/consultations/<consultation_id>/vitals/<vitals_id>',
                  methods=['DELETE'], strict_slashes=False)
 @swag_from('documentation/patients/patient_id/consultation/consultation_id/vitals/delete_vitals.yml',
            methods=['DELETE'])
-
-def delete_vitals(vitals_id):
+def delete_vitals(pid, consultation_id, vital_id):
     """
     Deletes a VitalSign Object for a specific consultation
     """
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
 
-    vitals = storage.get(VitalSign, vitals_id)
-    
+    consultation = storage.get('Consultation', 'id', consultation_id)
+    if (not consultation) or (consultation.pid != pid):
+        abort(404, description="Consultation not found")
+
+    vitals = storage.get('VitalSign', 'id', vital_id)
     if not vitals:
-        abort(404)
+        abort(404, description="No such vitals")
 
     storage.delete(vitals)
     storage.save()
@@ -63,18 +88,24 @@ def delete_vitals(vitals_id):
     return make_response(jsonify({}), 200)
 
 
-@app_views.route('/patients/<pid>/<consultation_id>',
+@app_views.route('/patients/<int:pid>/<consultation_id>',
                  methods=['POST'], strict_slashes=False)
 @swag_from('documentation/patient/patient_id/consultation_id/post_vitals.yml', methods=['POST'])
-def post_vitals(pid, consulation_id):
+def post_vitals(pid, consultation_id, vital_id):
     """
     Creates a new vitals for a specific consultation
     """
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
+
+    consultation = storage.get('Consultation', 'id', consultation_id)
+    if (not consultation) or (consultation.pid != pid):
+        abort(404, description="Consultation not found")
+
     if not request.get_json():
         abort(400, description="Not a JSON")
 
-    pid = pid
-    consulation_id = consulation_id
     if 'pr' not in request.get_json():
         abort(400, description="Missing pulse rate")
     if 'rr' not in request.get_json():
@@ -83,30 +114,35 @@ def post_vitals(pid, consulation_id):
         abort(400, description="Missing blood pressure")
     if 'temp' not in request.get_json():
         abort(400, description="Missing temperature")
-    if 'spo2' not in request.get_json():
-        abort(400, description="Missing oxygen saturation")
-    if 'height' not in request.get_json():
-        abort(400, description="Missing height")
-    if 'weight' not in request.get_json():
-        abort(400, description="Missing weight")
 
     data = request.get_json()
+    data['pid'] = pid
+    data['consultation_id'] = consultation_id
     instance = VitalSign(**data)
     instance.save()
+    instance = storage.get('VitalSign', 'id', instance.id)
     return make_response(jsonify(instance.to_dict()), 201)
 
-@app_views.route('/patients/<pid>/<consultation_id>/<vitals_id>', methods=['PUT'],
+
+@app_views.route('/patients/<int:pid>/<consultation_id>/<vitals_id>', methods=['PUT'],
                  strict_slashes=False)
 @swag_from('documentation/patient/patient_id/consultation_id/put_vitals.yml',
            methods=['PUT'])
-def put_vitals(vitals_id):
+def put_vitals(pid, consultation_id, vitals_id):
     """
     Updates a vitals for a specific consultation
     """
-    vitals = storage.get(VitalSign, vitals_id)
+    patient = storage.get('Patient', 'pid', pid)
+    if not patient:
+        abort(404, description="Patient not found")
 
+    consultation = storage.get('Consultation', 'id', consultation_id)
+    if (not consultation) or (consultation.pid != pid):
+        abort(404, description="Consultation not found")
+
+    vitals = storage.get('VitalSign', 'id', vitals_id)
     if not vitals:
-        abort(404)
+        abort(404, description='No such vitals')
 
     if not request.get_json():
         abort(400, description="Not a JSON")
@@ -117,5 +153,7 @@ def put_vitals(vitals_id):
     for key, value in data.items():
         if key not in ignore:
             setattr(vitals, key, value)
+
     storage.save()
+    vitals = storage.get('VitalSign', 'id', vitals.id)
     return make_response(jsonify(vitals.to_dict()), 200)
